login undefined issue resolve

+++ src/components/Login.jsx
@@ -33,8 +33,12 @@ const Login = () => {
       const userExist = checkUsers?.userData?.filter((user)=>{
         return user?.email === values?.email && user?.password === values?.password
       })
-      //match nai thay tyare undefined vali error ahiya thi aavse
-      localStorage.setItem('User', JSON.stringify(userExist[0]));
+
+      if(userExist?.length === 0){
+        localStorage.setItem('User', JSON.stringify([]));
+      }else{
+        localStorage.setItem('User', JSON.stringify(userExist[0]));
+      }
       formik?.resetForm();
       navigate('/')
     },
